ingest/requirements.txt
llama-index==0.10.65
llama-index-vector-stores-weaviate==0.1.10
llama-index-embeddings-ollama==0.3.1
weaviate-client==4.7.1
pandas==2.2.2
